Add files via upload
the signout button was labeled cancel so i just updated the label names and change the image name.

Started the parameter editing window and created and entire class for it. Right now its hard coded to only work on what ever the top left button is. msg me if you wanna know why cause its hard to explain put it into word.

The parameter window has and ok and apply button. Both will need to save the parameters on a file.